requirements.txt
streamlit
streamlit-option-menu
pillow
pandas
numpy
plotly
seaborn
matplotlib
scikit-learn
imbalanced-learn
scipy
